Selecting Time State Created
Once a provider views a schedule that has available times, they will
choose 1 time/table cell. Add class "selected" to the td and the cell
will turn blue. For this class to work there must be class
"highlighted" already on the call, which will be there from my
scheduling js.