Homepage w/ issues
Can't figure out how to make the waypoints plugin fire when using
Bootstraps carousel at the top of the homepage. I need the waypoints
plugin to work so that I can create animations based on scroll
events…for example adding classes to elements when the user scrolls to
a certain section of the page.